Complete rewrite of server commands processing.

+++ rules.h
@@ -1,17 +1,19 @@
 #ifndef RULES_H
 #define RULES_H
 
+#include "types.h"
+
 typedef struct rules 
 {
-	int loadtime;
-	int turntime;
-	int rows;
-	int cols;
-	int turns;
-	int viewradius_sq;
-	int attackradius_sq;
-	int spawnradius_sq;
-	int seed;
+	guint loadtime;
+	guint turntime;
+	guint rows;
+	guint cols;
+	guint turns;
+	guint viewradius_sq;
+	guint attackradius_sq;
+	guint spawnradius_sq;
+	gint64 seed;
 } Rules;
 
 #endif // RULES_H
